+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Handle `TransportError` and show nice error message
 
 ### Changed
 - [Dev] Simplify packaging setup (fix #284)